server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
clarifai==9.8.2
clarifai-grpc==9.8.4
click==8.1.7
cmake==3.27.5
cohere==4.27
contextlib2==21.6.0
cssselect==1.2.0
dataclasses-json==0.6.1
dill==0.3.7
exceptiongroup==1.1.3
fastapi==0.103.1
fastavro==1.8.2
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.2
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.58.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.302
langsmith==0.0.40
lit==17.0.1
lxml==4.9.3
manifest-ml==0.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
newspaper3k==0.2.8
nlpcloud==1.1.44
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.28.1
openlm==0.0.5
packaging==23.1
Pillow==10.0.1
protobuf==4.24.3
pydantic==2.4.1
pydantic_core==2.10.1
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-rapidjson==1.11
PyYAML==6.0.1
redis==5.0.1
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
rich==13.4.2
safetensors==0.3.3
schema==0.7.5
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
sqlitedict==2.1.0
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tinysegmenter==0.3
tldextract==3.6.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.64.1
transformers==4.33.2
triton==2.0.0
tritonclient==2.34.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.5
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
Whoosh==2.7.4
yarl==1.9.2
zipp==3.17.0
